+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 **Unreleased**
 --------------
 
+- **Enhancement**: Report more context for error types, such as the name of the parameter it came from.
+- **Fix**: For generic types, check for error types in their type arguments as well.
+
 0.3.1
 -----